+++ packages/astro/src/core/preview/index.ts
@@ -1,4 +1,4 @@
-import type { AstroConfig } from '../../@types/astro-core';
+import type { AstroConfig } from '../../@types/astro';
 import type { LogOptions } from '../logger';
 
 import http from 'http';
@@ -23,7 +23,7 @@ interface PreviewServer {
 /** The primary dev action */
 export default async function preview(config: AstroConfig, { logging }: PreviewOptions): Promise<PreviewServer> {
   const startServerTime = performance.now();
-  const base = config.buildOptions.site ? new URL(config.buildOptions.site).pathname + '/' : '/';
+  const base = config.buildOptions.site ? new URL(config.buildOptions.site).pathname : '/';
 
   // Create the preview server, send static files out of the `dist/` directory.
   const server = http.createServer((req, res) => {