+++ packages/astro/src/runtime/server/index.ts
@@ -1,6 +1,5 @@
-import type { AstroComponentMetadata, Renderer } from '../../@types/astro-core';
-import type { SSRResult, SSRElement } from '../../@types/astro-runtime';
-import type { TopLevelAstro } from '../../@types/astro-runtime';
+import type { AstroComponentMetadata, Renderer } from '../../@types/astro';
+import type { AstroGlobalPartial, SSRResult, SSRElement } from '../../@types/astro';
 
 import shorthash from 'shorthash';
 import { extractDirectives, generateHydrateScript } from './hydration.js';
@@ -18,7 +17,7 @@ export type { Metadata } from './metadata';
 async function _render(child: any): Promise<any> {
   child = await child;
   if (Array.isArray(child)) {
-    return (await Promise.all(child.map((value) => _render(value)))).join('\n');
+    return (await Promise.all(child.map((value) => _render(value)))).join('');
   } else if (typeof child === 'function') {
     // Special: If a child is a function, call it automatically.
     // This lets you do {() => ...} without the extra boilerplate
@@ -31,7 +30,7 @@ async function _render(child: any): Promise<any> {
   }
   // Add a comment explaining why each of these are needed.
   // Maybe create clearly named function for what this is doing.
-  else if (child instanceof AstroComponent || child.toString() === '[object AstroComponent]') {
+  else if (child instanceof AstroComponent || Object.prototype.toString.call(child) === '[object AstroComponent]') {
     return await renderAstroComponent(child);
   } else {
     return child;
@@ -93,10 +92,31 @@ export async function renderSlot(_result: any, slotted: string, fallback?: any) 
 
 export const Fragment = Symbol('Astro.Fragment');
 
+function guessRenderers(componentUrl?: string): string[] {
+  const extname = componentUrl?.split('.').pop();
+  switch (extname) {
+    case 'svelte':
+      return ['@astrojs/renderer-svelte'];
+    case 'vue':
+      return ['@astrojs/renderer-vue'];
+    case 'jsx':
+    case 'tsx':
+      return ['@astrojs/renderer-react', '@astrojs/renderer-preact'];
+    default:
+      return ['@astrojs/renderer-react', '@astrojs/renderer-preact', '@astrojs/renderer-vue', '@astrojs/renderer-svelte'];
+  }
+}
+
+function formatList(values: string[]): string {
+  if (values.length === 1) {
+    return values[0];
+  }
+  return `${values.slice(0, -1).join(', ')} or ${values[values.length - 1]}`;
+}
+
 export async function renderComponent(result: SSRResult, displayName: string, Component: unknown, _props: Record<string | number, any>, slots: any = {}) {
   Component = await Component;
   const children = await renderSlot(result, slots?.default);
-  const { renderers } = result._metadata;
 
   if (Component === Fragment) {
     return children;
@@ -107,11 +127,12 @@ export async function renderComponent(result: SSRResult, displayName: string, Co
     return output;
   }
 
-  let metadata: AstroComponentMetadata = { displayName };
-
-  if (Component == null) {
-    throw new Error(`Unable to render ${metadata.displayName} because it is ${Component}!\nDid you forget to import the component or is it possible there is a typo?`);
+  if (Component === null && !_props['client:only']) {
+    throw new Error(`Unable to render ${displayName} because it is ${Component}!\nDid you forget to import the component or is it possible there is a typo?`);
   }
+
+  const { renderers } = result._metadata;
+  const metadata: AstroComponentMetadata = { displayName };
 
   const { hydration, props } = extractDirectives(_props);
   let html = '';
@@ -122,33 +143,93 @@ export async function renderComponent(result: SSRResult, displayName: string, Co
     metadata.componentExport = hydration.componentExport;
     metadata.componentUrl = hydration.componentUrl;
   }
+  const probableRendererNames = guessRenderers(metadata.componentUrl);
+
+  if (Array.isArray(renderers) && renderers.length === 0 && typeof Component !== 'string') {
+    const message = `Unable to render ${metadata.displayName}! 
+
+There are no \`renderers\` set in your \`astro.config.mjs\` file.
+Did you mean to enable ${formatList(probableRendererNames.map((r) => '`' + r + '`'))}?`;
+    throw new Error(message);
+  }
 
   // Call the renderers `check` hook to see if any claim this component.
   let renderer: Renderer | undefined;
-  for (const r of renderers) {
-    if (await r.ssr.check(Component, props, children)) {
-      renderer = r;
-      break;
+  if (metadata.hydrate !== 'only') {
+    for (const r of renderers) {
+      if (await r.ssr.check(Component, props, children)) {
+        renderer = r;
+        break;
+      }
+    }
+  } else {
+    // Attempt: use explicitly passed renderer name
+    if (metadata.hydrateArgs) {
+      const rendererName = metadata.hydrateArgs;
+      renderer = renderers.filter(({ name }) => name === `@astrojs/renderer-${rendererName}` || name === rendererName)[0];
+    }
+    // Attempt: can we guess the renderer from the export extension?
+    if (!renderer) {
+      const extname = metadata.componentUrl?.split('.').pop();
+      renderer = renderers.filter(({ name }) => name === `@astrojs/renderer-${extname}` || name === extname)[0];
     }
   }
 
   // If no one claimed the renderer
   if (!renderer) {
-    // This is a custom element without a renderer. Because of that, render it
-    // as a string and the user is responsible for adding a script tag for the component definition.
-    if (typeof Component === 'string') {
-      html = await renderAstroComponent(await render`<${Component}${spreadAttributes(props)}>${children}</${Component}>`);
-    } else {
-      throw new Error(`Astro is unable to render ${metadata.displayName}!\nIs there a renderer to handle this type of component defined in your Astro config?`);
+    if (metadata.hydrate === 'only') {
+      // TODO: improve error message
+      throw new Error(`Unable to render ${metadata.displayName}!
+
+Using the \`client:only\` hydration strategy, Astro needs a hint to use the correct renderer.
+Did you mean to pass <${metadata.displayName} client:only="${probableRendererNames.map((r) => r.replace('@astrojs/renderer-', '')).join('|')}" />
+`);
+    } else if (typeof Component !== 'string') {
+      const matchingRenderers = renderers.filter((r) => probableRendererNames.includes(r.name));
+      const plural = renderers.length > 1;
+      if (matchingRenderers.length === 0) {
+        throw new Error(`Unable to render ${metadata.displayName}!
+
+There ${plural ? 'are' : 'is'} ${renderers.length} renderer${plural ? 's' : ''} configured in your \`astro.config.mjs\` file,
+but ${plural ? 'none were' : 'it was not'} able to server-side render ${metadata.displayName}.
+
+Did you mean to enable ${formatList(probableRendererNames.map((r) => '`' + r + '`'))}?`);
+      } else {
+        throw new Error(`Unable to render ${metadata.displayName}!
+
+This component likely uses ${formatList(probableRendererNames)},
+but Astro encountered an error during server-side rendering.
+
+Please ensure that ${metadata.displayName}:
+1. Does not unconditionally access browser-specific globals like \`window\` or \`document\`.
+   If this is unavoidable, use the \`client:only\` hydration directive.
+2. Does not conditionally return \`null\` or \`undefined\` when rendered on the server.
+
+If you're still stuck, please open an issue on GitHub or join us at https://astro.build/chat.`);
+      }
     }
   } else {
-    ({ html } = await renderer.ssr.renderToStaticMarkup(Component, props, children));
+    if (metadata.hydrate === 'only') {
+      html = await renderSlot(result, slots?.fallback);
+    } else {
+      ({ html } = await renderer.ssr.renderToStaticMarkup(Component, props, children));
+    }
+  }
+
+  // This is a custom element without a renderer. Because of that, render it
+  // as a string and the user is responsible for adding a script tag for the component definition.
+  if (!html && typeof Component === 'string') {
+    html = await renderAstroComponent(await render`<${Component}${spreadAttributes(props)}>${children}</${Component}>`);
   }
 
   // This is used to add polyfill scripts to the page, if the renderer needs them.
   if (renderer?.polyfills?.length) {
-    let polyfillScripts = renderer.polyfills.map((src) => `<script type="module">import "${src}";</script>`).join('');
-    html = html + polyfillScripts;
+    for (const src of renderer.polyfills) {
+      result.scripts.add({
+        props: { type: 'module' },
+        children: `import "${src}";`,
+      });
+    }
   }
 
   if (!hydration) {
@@ -162,7 +243,7 @@ export async function renderComponent(result: SSRResult, displayName: string, Co
   // INVESTIGATE: This will likely be a problem in streaming because the `<head>` will be gone at this point.
   result.scripts.add(await generateHydrateScript({ renderer, astroId, props }, metadata as Required<AstroComponentMetadata>));
 
-  return `<astro-root uid="${astroId}">${html}</astro-root>`;
+  return `<astro-root uid="${astroId}">${html ?? ''}</astro-root>`;
 }
 
 /** Create the Astro.fetchContent() runtime function. */
@@ -181,6 +262,7 @@ function createFetchContentFn(url: URL) {
         const urlSpec = new URL(spec, url).pathname;
         return {
           ...mod.frontmatter,
+          Content: mod.default,
           content: mod.metadata,
           file: new URL(spec, url),
           url: urlSpec.includes('/pages/') ? urlSpec.replace(/^.*\/pages\//, '/').replace(/(\/index)?\.md$/, '') : undefined,
@@ -191,12 +273,12 @@ function createFetchContentFn(url: URL) {
   // This has to be cast because the type of fetchContent is the type of the function
   // that receives the import.meta.glob result, but the user is using it as
   // another type.
-  return fetchContent as unknown as TopLevelAstro['fetchContent'];
+  return fetchContent as unknown as AstroGlobalPartial['fetchContent'];
 }
 
 // This is used to create the top-level Astro global; the one that you can use
 // Inside of getStaticPaths.
-export function createAstro(fileURLStr: string, site: string): TopLevelAstro {
+export function createAstro(fileURLStr: string, site: string): AstroGlobalPartial {
   const url = new URL(fileURLStr);
   const fetchContent = createFetchContentFn(url);
   return {