+++ packages/astro/src/vite-plugin-build-css/index.ts
@@ -1,10 +1,9 @@
-import type { ResolveIdHook, LoadHook, RenderedChunk } from 'rollup';
-import type { Plugin as VitePlugin } from 'vite';
+import type { RenderedChunk } from 'rollup';
+import { Plugin as VitePlugin } from '../core/vite';
 
 import { STYLE_EXTENSIONS } from '../core/ssr/css.js';
-import { getViteResolve, getViteLoad } from './resolve.js';
-import { getViteTransform, TransformHook } from '../vite-plugin-astro/styles.js';
 import * as path from 'path';
+import esbuild from 'esbuild';
 
 const PLUGIN_NAME = '@astrojs/rollup-plugin-build-css';
 
@@ -56,15 +55,31 @@ export function rollupPluginAstroBuildCSS(options: PluginOptions): VitePlugin {
   const { astroPageStyleMap, astroStyleMap, chunkToReferenceIdMap, pureCSSChunks } = options;
   const styleSourceMap = new Map<string, string>();
 
-  let viteTransform: TransformHook;
-
   return {
     name: PLUGIN_NAME,
 
-    enforce: 'pre',
-
     configResolved(resolvedConfig) {
-      viteTransform = getViteTransform(resolvedConfig);
+      // Our plugin needs to run before `vite:css-post` which does a lot of what we do
+      // for bundling CSS, but since we need to control CSS we should go first.
+      // We move to right before the vite:css-post plugin so that things like the
+      // Vue plugin go before us.
+      const plugins = resolvedConfig.plugins as VitePlugin[];
+      const viteCSSPostIndex = resolvedConfig.plugins.findIndex((p) => p.name === 'vite:css-post');
+      if (viteCSSPostIndex !== -1) {
+        const viteCSSPost = plugins[viteCSSPostIndex];
+        // Prevent this plugin's bundling behavior from running since we need to
+        // do that ourselves in order to handle updating the HTML.
+        delete viteCSSPost.renderChunk;
+        delete viteCSSPost.generateBundle;
+
+        // Move our plugin to be right before this one.
+        const ourIndex = plugins.findIndex((p) => p.name === PLUGIN_NAME);
+        const ourPlugin = plugins[ourIndex];
+
+        // Remove us from where we are now and place us right before the viteCSSPost plugin
+        plugins.splice(ourIndex, 1);
+        plugins.splice(viteCSSPostIndex - 1, 0, ourPlugin);
+      }
     },
 
     async resolveId(id) {
@@ -93,21 +108,15 @@ export function rollupPluginAstroBuildCSS(options: PluginOptions): VitePlugin {
         styleSourceMap.set(id, value);
         return null;
       }
-      if (isCSSRequest(id)) {
-        let result = await viteTransform(value, id);
-        if (result) {
-          styleSourceMap.set(id, result.code);
-        } else {
-          styleSourceMap.set(id, value);
-        }
 
-        return result;
+      if (isCSSRequest(id)) {
+        styleSourceMap.set(id, value);
       }
 
       return null;
     },
 
-    renderChunk(_code, chunk) {
+    async renderChunk(_code, chunk) {
       let chunkCSS = '';
       let isPureCSS = true;
       for (const [id] of Object.entries(chunk.modules)) {
@@ -120,10 +129,14 @@ export function rollupPluginAstroBuildCSS(options: PluginOptions): VitePlugin {
       }
 
       if (isPureCSS) {
+        const { code: minifiedCSS } = await esbuild.transform(chunkCSS, {
+          loader: 'css',
+          minify: true,
+        });
         const referenceId = this.emitFile({
           name: chunk.name + '.css',
           type: 'asset',
-          source: chunkCSS,
+          source: minifiedCSS,
         });
         pureCSSChunks.add(chunk);
         chunkToReferenceIdMap.set(chunk.fileName, referenceId);
@@ -133,10 +146,31 @@ export function rollupPluginAstroBuildCSS(options: PluginOptions): VitePlugin {
     },
 
     // Delete CSS chunks so JS is not produced for them.
-    generateBundle(_options, bundle) {
-      for (const [chunkId, chunk] of Object.entries(bundle)) {
-        if (chunk.type === 'chunk' && pureCSSChunks.has(chunk)) {
-          delete bundle[chunkId];
+    generateBundle(opts, bundle) {
+      if (pureCSSChunks.size) {
+        const pureChunkFilenames = new Set([...pureCSSChunks].map((chunk) => chunk.fileName));
+        const emptyChunkFiles = [...pureChunkFilenames]
+          .map((file) => path.basename(file))
+          .join('|')
+          .replace(/\./g, '\\.');
+        const emptyChunkRE = new RegExp(opts.format === 'es' ? `\\bimport\\s*"[^"]*(?:${emptyChunkFiles})";\n?` : `\\brequire\\(\\s*"[^"]*(?:${emptyChunkFiles})"\\);\n?`, 'g');
+
+        for (const [chunkId, chunk] of Object.entries(bundle)) {
+          if (chunk.type === 'chunk') {
+            if (pureCSSChunks.has(chunk)) {
+              // Delete pure CSS chunks, these are JavaScript chunks that only import
+              // other CSS files, so are empty at the end of bundling.
+              delete bundle[chunkId];
+            } else {
+              // Remove any pure css chunk imports from JavaScript.
+              // Note that this code comes from Vite's CSS build plugin.
+              chunk.code = chunk.code.replace(
+                emptyChunkRE,
+                // remove css import while preserving source map location
+                (m) => `/* empty css ${''.padEnd(m.length - 15)}*/`
+              );
+            }
+          }
         }
       }
     },