+++ packages/astro/src/vite-plugin-build-html/index.ts
@@ -1,6 +1,6 @@
-import type { AstroConfig, RouteCache } from '../@types/astro-core';
+import type { AstroConfig, RouteCache } from '../@types/astro';
 import type { LogOptions } from '../core/logger';
-import type { ViteDevServer, Plugin as VitePlugin } from 'vite';
+import type { ViteDevServer, Plugin as VitePlugin } from '../core/vite';
 import type { OutputChunk, PreRenderedChunk, RenderedChunk } from 'rollup';
 import type { AllPagesData } from '../core/build/types';
 import parse5 from 'parse5';
@@ -12,7 +12,6 @@ import { addRollupInput } from './add-rollup-input.js';
 import { findAssets, findExternalScripts, findInlineScripts, findInlineStyles, getTextContent, isStylesheetLink } from './extract-assets.js';
 import { render as ssrRender } from '../core/ssr/index.js';
 import { getAstroStyleId, getAstroPageStyleId } from '../vite-plugin-build-css/index.js';
-import { viteifyPath } from '../core/util.js';
 
 // This package isn't real ESM, so have to coerce it
 const matchSrcset: typeof srcsetParse = (srcsetParse as any).default;
@@ -26,8 +25,14 @@ const ASTRO_EMPTY = '@astro-empty';
 const tagsWithSrcSet = new Set(['img', 'source']);
 
 const isAstroInjectedLink = (node: parse5.Element) => isStylesheetLink(node) && getAttribute(node, 'data-astro-injected') === '';
-const isBuildableLink = (node: parse5.Element, srcRoot: string) => isAstroInjectedLink(node) || getAttribute(node, 'href')?.startsWith(srcRoot);
-const isBuildableImage = (node: parse5.Element, srcRoot: string) => getTagName(node) === 'img' && getAttribute(node, 'src')?.startsWith(srcRoot);
+const isBuildableLink = (node: parse5.Element, srcRoot: string, srcRootWeb: string) => {
+  if (isAstroInjectedLink(node)) return true;
+  const href = getAttribute(node, 'href');
+  if (typeof href !== 'string' || !href.length) return false;
+  return href.startsWith(srcRoot) || href.startsWith(srcRootWeb) || `/${href}`.startsWith(srcRoot); // Windows fix: some paths are missing leading "/"
+};
+const isBuildableImage = (node: parse5.Element, srcRoot: string, srcRootWeb: string) =>
+  getTagName(node) === 'img' && (getAttribute(node, 'src')?.startsWith(srcRoot) || getAttribute(node, 'src')?.startsWith(srcRootWeb));
 const hasSrcSet = (node: parse5.Element) => tagsWithSrcSet.has(getTagName(node)) && !!getAttribute(node, 'srcset');
 const isHoistedScript = (node: parse5.Element) => getTagName(node) === 'script' && hasAttribute(node, 'hoist');
 
@@ -48,7 +53,10 @@ interface PluginOptions {
 export function rollupPluginAstroBuildHTML(options: PluginOptions): VitePlugin {
   const { astroConfig, astroStyleMap, astroPageStyleMap, chunkToReferenceIdMap, pureCSSChunks, logging, origin, allPages, routeCache, viteServer, pageNames } = options;
 
+  // The filepath root of the src folder
   const srcRoot = astroConfig.src.pathname;
+  // The web path of the src folter
+  const srcRootWeb = srcRoot.substr(astroConfig.projectRoot.pathname.length - 1);
 
   // A map of pages to rendered HTML
   const renderedPageMap = new Map<string, string>();
@@ -131,13 +139,12 @@ export function rollupPluginAstroBuildHTML(options: PluginOptions): VitePlugin {
 
           const assetImports = [];
           for (let node of findAssets(document)) {
-            if (isBuildableLink(node, srcRoot)) {
+            if (isBuildableLink(node, srcRoot, srcRootWeb)) {
               const href = getAttribute(node, 'href')!;
-              const linkId = viteifyPath(href);
-              assetImports.push(linkId);
+              assetImports.push(href);
             }
 
-            if (isBuildableImage(node, srcRoot)) {
+            if (isBuildableImage(node, srcRoot, srcRootWeb)) {
               const src = getAttribute(node, 'src');
               if (src?.startsWith(srcRoot) && !astroAssetMap.has(src)) {
                 astroAssetMap.set(src, fs.readFile(new URL(`file://${src}`)));
@@ -353,7 +360,7 @@ export function rollupPluginAstroBuildHTML(options: PluginOptions): VitePlugin {
         const styleId = getAstroPageStyleId(pathname);
         let pageCSSAdded = false;
         for (const node of findAssets(document)) {
-          if (isBuildableLink(node, srcRoot)) {
+          if (isBuildableLink(node, srcRoot, srcRootWeb)) {
             const rel = getAttribute(node, 'rel');
             switch (rel) {
               case 'stylesheet': {
@@ -375,7 +382,7 @@ export function rollupPluginAstroBuildHTML(options: PluginOptions): VitePlugin {
             }
           }
 
-          if (isBuildableImage(node, srcRoot)) {
+          if (isBuildableImage(node, srcRoot, srcRootWeb)) {
             const src = getAttribute(node, 'src')!;
             const referenceId = assetIdMap.get(src);
             if (referenceId) {