+++ packages/astro/test/fixtures/errors/src/components/SolidSyntaxError.jsx
@@ -1,0 +1,5 @@
+import { h } from 'solid-js/web'
+
+export default function ReactSyntaxError() {
+  return (<div></div></div>);
+}