+++ packages/astro/test/fixtures/postcss/src/components/Vue.vue
@@ -1,0 +1,12 @@
+<style>
+.vue {
+  appearance: none;
+}
+</style>
+
+<template>
+  <div class="vue">
+    Vue
+  </div>
+</template>
+