+++ packages/markdown/remark/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @astrojs/markdown-support
 
+## 0.4.0
+
+### Minor Changes
+
+- e6aaeff5: Initial release.
+
+### Patch Changes
+
+- Updated dependencies [e6aaeff5]
+  - @astrojs/prism@0.3.0
+
+## 0.4.0-next.2
+
+### Patch Changes
+
+- 00d2b625: Move gray-matter to deps
+
 ## 0.4.0-next.1
 
 ### Patch Changes